pyproject.toml
[tool.poetry]
name = "course-enrollment"
version = "0.1.0"
description = ""
authors = ["sudomanoj <manoj.765421@bumc.tu.edu.np>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
django-jazzmin = "^3.0.1"
django-environ = "^0.12.0"
django-nested-admin = "^4.1.1"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
